pivotal-181944040 refactor
